Add some error handling for delete endpoint project

+++ src/main/java/nl/mpdev/project_manager_backend/services/ProjectService.java
@@ -38,6 +38,9 @@ public class ProjectService {
   } 
 
   public void deleteProject(Long id){
+    if(!projectRepository.existsById(id)){
+      throw new RuntimeException("Record was not found");
+    }
     projectRepository.deleteById(id);
   }
 }
